Refactor: UploadImageContainer 리팩토링 및 프로필 수정 페이지 내 프로필 사진 변경 컴포넌트로 대체 (#152)
