+++ src/containers/common/UploadImageContainer.jsx
@@ -1,20 +1,27 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import UploadImage from '../../components/common/UploadImage';
-import { initImage, uploadImage } from '../../modules/image';
+import { initImage, uploadProfileImage } from '../../modules/image';
 
 const UploadImageContainer = (props) => {
-  const { images } = useSelector((state) => state.image);
+  const { profileImage } = useSelector((state) => state.image);
   const dispatch = useDispatch();
 
-  const onUploadImage = (formData) => {
-    dispatch(uploadImage(formData));
+  const onUploadProfileImage = (formData) => {
+    dispatch(uploadProfileImage(formData));
   };
 
   const onInitImage = () => {
     dispatch(initImage());
   };
-  return <UploadImage images={images} onUploadImage={onUploadImage} onInitImage={onInitImage} />;
+
+  return (
+    <UploadImage
+      profileImage={profileImage}
+      onUploadProfileImage={onUploadProfileImage}
+      onInitImage={onInitImage}
+    />
+  );
 };
 
 export default UploadImageContainer;
