+++ src/containers/profile/setting/ProfileSettingContainer.jsx
@@ -1,18 +1,18 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { initImage, setProfileImage, uploadProfileImage } from '../../../modules/image';
 import { check } from '../../../modules/user';
 import { updateProfile } from '../../../modules/write';
 import { checkNickname, initAuth } from '../../../modules/auth';
 import ProfileSetting from '../../../components/profile/setting/ProfileSetting';
 import { changePassword, checkPassword } from '../../../modules/profile';
+import { setProfileImage } from '../../../modules/image';
 
 const ProfileSettingContainer = () => {
   const dispatch = useDispatch();
   const { user } = useSelector((state) => state.user);
   const { auth, nicknameChecked } = useSelector((state) => state.auth);
-  const { updatedProfile } = useSelector((state) => state.write);
   const { profileImage } = useSelector((state) => state.image);
+  const { updatedProfile } = useSelector((state) => state.write);
   const { checkedPassword, changedPassword } = useSelector((state) => state.profile);
 
   const onChangePassword = ({ password }) => {
@@ -37,27 +37,12 @@ const ProfileSettingContainer = () => {
     );
   };
 
-  const onUploadProfileImage = (formData) => {
-    dispatch(uploadProfileImage(formData));
-  };
-
-  const onInitImage = () => {
-    dispatch(initImage());
-  };
-
   const onCheckNickname = ({ nickname }) => {
     dispatch(checkNickname({ nickname }));
   };
 
   useEffect(() => {
-    if (user?.profile?.profileImage) {
-      dispatch(setProfileImage(user.profile.profileImage));
-    }
-  }, [user]);
-
-  useEffect(() => {
     return () => {
-      dispatch(initImage());
       dispatch(initAuth());
     };
   }, [dispatch]);
@@ -68,12 +53,13 @@ const ProfileSettingContainer = () => {
     }
   }, [auth, dispatch]);
 
+  useEffect(() => {
+    dispatch(setProfileImage(user?.profile?.profileImage));
+  }, [dispatch, user]);
+
   return (
     <ProfileSetting
       user={user}
-      profileImage={profileImage}
-      onUploadProfileImage={onUploadProfileImage}
-      onInitImage={onInitImage}
       onCheckNickname={onCheckNickname}
       nicknameDuplicateError={!nicknameChecked}
       onUpdateProfile={onUpdateProfile}
